separate out data util tests into separate files and their own folder for better organization
