+++ data/src/test/data_utils.py/test_contiguous_neighbors.py
@@ -1,0 +1,15 @@
+import unittest
+
+import geopandas as gpd
+
+from new_etl.classes.featurelayer import FeatureLayer
+from new_etl.data_utils.contig_neighbors import contig_neighbors
+
+
+class TestContiguousNeighbors(unittest.TestCase):
+    def test_contiguous_neighbors(self):
+        pass
+
+
+if __name__ == "__main__":
+    unittest.main()
