+++ data/src/test/test_data_utils.py
@@ -4,26 +4,12 @@ from io import BytesIO
 from unittest.mock import MagicMock, Mock, patch
 
 import geopandas as gpd
-import numpy as np
 from shapely.geometry import LineString, MultiPolygon, Point, Polygon
 
 from config.config import USE_CRS
 from data_utils.park_priority import get_latest_shapefile_url, park_priority
 from data_utils.ppr_properties import ppr_properties
 from data_utils.vacant_properties import vacant_properties
-from new_etl.classes.featurelayer import FeatureLayer
-from new_etl.data_utils import (
-    access_process,
-    conservatorship,
-    negligent_devs,
-    owner_type,
-    priority_level,
-    tactical_urbanism,
-)
-from new_etl.data_utils.pwd_parcels import (
-    merge_pwd_parcels_gdf,
-    transform_pwd_parcels_gdf,
-)
 
 
 class TestDataUtils(unittest.TestCase):
@@ -176,287 +162,6 @@ class TestDataUtils(unittest.TestCase):
         Test the vacant properties layer. Simply construct the class to see if it works.
         """
         vacant_properties()
-
-    def test_pwd_parcels_merge(self):
-        """
-        This tests that the merge_pwd_parcels_gdf function correctly retains
-        existing point geometries in the primary GeoDataFrame when no better
-        geometry is available in the PWD parcels GeoDataFrame.
-        """
-
-        parcel_type_sample_data = ["Land", "Building", "Land", "Building"]
-        primary_data = {
-            "opa_id": ["0100", "0101", "0102", "0103"],
-            "geometry": [Point(0, 0), Point(1, 1), Point(2, 2), Point(3, 3)],
-            "parcel_type": parcel_type_sample_data,
-        }
-
-        # CRS we don't use is chosen since we are just checking it doesn't change:
-        test_crs = "EPSG:2002"
-        primary_gdf = gpd.GeoDataFrame(primary_data, geometry="geometry", crs=test_crs)
-
-        # Simulate dropped rows by having a gapped index.
-        primary_gdf.index = [0, 2, 3, 5]
-
-        # Create pwd_parcels_gdf with a continuous index.
-        pwd_data = {
-            "opa_id": ["0100", "0101", "0102", "0103"],
-            "geometry": [
-                np.nan,  # For opa 0100, no better geometry
-                MultiPolygon(
-                    [Polygon([(10, 10), (20, 10), (20, 20), (10, 20)])]
-                ),  # For opa 0101, a multipolygon available
-                np.nan,
-                np.nan,
-            ],
-        }
-        pwd_gdf = gpd.GeoDataFrame(pwd_data, geometry="geometry")
-        assert list(pwd_gdf.index) == list(range(4))  # Continuous index: 0,1,2,3
-
-        # Call the merge function under test.
-        merged_gdf = merge_pwd_parcels_gdf(primary_gdf, pwd_gdf)
-
-        # Expected geometries:
-        # For opa_id "0010", we expect the multipolygon from pwd_gdf.
-        # For the other opa_id's we expect the point geometries from primary_gdf.
-        expected_geometries = {
-            "0100": primary_gdf.loc[0, "geometry"],
-            "0101": pwd_gdf.loc[1, "geometry"],  # pwd geometry used when available
-            "0102": primary_gdf.loc[3, "geometry"],
-            "0103": primary_gdf.loc[5, "geometry"],
-        }
-
-        assert merged_gdf.crs == primary_gdf.crs
-        assert merged_gdf.crs == test_crs
-
-        # Verify each row of the merged GeoDataFrame.
-        for idx, row in merged_gdf.iterrows():
-            opa_id = row["opa_id"]
-            actual_geom = row["geometry"]
-            expected_geom = expected_geometries[opa_id]
-            # Use shapely's equals() to check geometry equivalence.
-            assert actual_geom.equals(expected_geom), (
-                f"Mismatch for opa_id {opa_id}: "
-                f"expected {expected_geom}, got {actual_geom}"
-            )
-
-        opa_id_values = ["0100", "0101", "0102", "0103"]
-        data = {
-            "opa_id": opa_id_values,
-            "geometry": [expected_geometries[opa_id] for opa_id in opa_id_values],
-            "parcel_type": parcel_type_sample_data,
-        }
-        expected_df = gpd.GeoDataFrame(
-            data=data,
-            geometry="geometry",
-        )
-
-        assert expected_df.equals(merged_gdf)
-
-    def test_transform_pwd_parcels_gdf_basic(self):
-        """
-        Check the basic functionality of the transform_pwd_parcels_gdf function.
-        This function is expected to mutate the input GeoDataFrame in place.
-        It should drop rows with null 'brt_id', rename 'brt_id' to 'opa_id',
-        and ensure all geometries are valid polygons or multipolygons.
-        """
-        # Create test data
-        data = {
-            "brt_id": [None, "1234", "5678"],
-            "geometry": [
-                Polygon(
-                    [(0, 0), (1, 0), (1, 1), (0, 1)]
-                ),  # Will be dropped (null brt_id)
-                Polygon([(0, 0), (2, 0), (2, 2), (0, 2)]),  # Valid polygon
-                LineString([(0, 0), (1, 1)]),  # Invalid type (not a polygon)
-            ],
-        }
-        gdf = gpd.GeoDataFrame()
-        gdf = gpd.GeoDataFrame(data, geometry="geometry")
-
-        # Expect a ValueError because LineString is not allowed
-        with self.assertRaises(ValueError) as context:
-            transform_pwd_parcels_gdf(gdf.copy())
-
-        self.assertIn("not polygons or multipolygons", str(context.exception))
-
-        # Now fix the invalid geometry to test normal flow
-        gdf.loc[2, "geometry"] = Polygon([(0, 0), (1, 0), (1, 1), (0, 1)])
-
-        # Run transformation which mutates result in place
-        result = gdf.copy()
-        transform_pwd_parcels_gdf(result)
-
-        # Check the rows were filtered
-        self.assertEqual(len(result), 2)  # One row was dropped
-        self.assertNotIn("brt_id", result.columns)
-        self.assertIn("opa_id", result.columns)
-        self.assertListEqual(sorted(result["opa_id"].tolist()), ["1234", "5678"])
-        self.assertTrue(all(result.geometry.is_valid))
-        self.assertTrue(all(result.geometry.type.isin(["Polygon", "MultiPolygon"])))
-
-    def test_owner_type(self):
-        """
-        Test the functionality of the owner_type service that generates a new column for the dataset
-        on the basis of existing ownership and opa_id data in the corresponding row.
-        """
-        data = {
-            "owner_1": ["John Smith", None, "Costco llc", "Jeff Roe"],
-            "owner_2": ["Jane Doe", "Tom Hink", "Sixers", "Jefferson llc"],
-            "city_owner_agency": [None, "City of Philadelphia", "PRA", None],
-        }
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-
-        owner_type_feature_layer = owner_type(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["owner_type"] = [
-            "Individual",
-            "Public",
-            "Public",
-            "Business (LLC)",
-        ]
-
-        assert expected_gdf.equals(owner_type_feature_layer.gdf)
-
-    def test_contiguous_neighbors(self):
-        pass
-
-    def test_negligent_devs(self):
-        """
-        Test the functionality of the negligent_devs service that generates new columns on the dataset relating to the number of properties owned
-        and vacant properties owned under one standardized address as well as a classification of the owner type as negligent based on the vacant number.
-        """
-        data = {
-            "opa_id": ["0011", "0011", "0011", "0011", "0011", "0000", "0000", "0032"],
-            "vacant": [True, True, True, True, True, False, True, True],
-            "city_owner_agency": [
-                None,
-                None,
-                None,
-                None,
-                None,
-                "City of Philadelphia",
-                "City of Philadelphia",
-                "PRA",
-            ],
-            "standardized_address": ["1", "2", "3", "4", "5", "6", "7", "8"],
-        }
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-        # need to add primary merge data
-        negligent_devs_feature_layer = negligent_devs(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["n_total_properties_owned"] = [1, 2, 3, 4]
-        expected_gdf["n_vacant_properties_owned"] = [1, 2, 3, 4]
-        expected_gdf["negligent_dev"] = [False, False, False, False]
-
-        assert expected_gdf.equals(negligent_devs_feature_layer.gdf)
-
-    def test_tactical_urbanism(self):
-        data = {
-            "parcel_type": ["Land", "Building", "Land", None],
-            "unsafe_building": ["N", "Y", "N", "N"],
-            "imm_dang_building": ["Y", "N", "N", "Y"],
-        }
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-
-        tactical_urbanism_feature_layer = tactical_urbanism(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["tactical_urbanism"] = ["No", "Yes", "Yes", "Yes"]
-
-        assert expected_gdf.equals(tactical_urbanism_feature_layer.gdf)
-
-    def test_conservatorship(self):
-        data = {
-            "city_owner_agency": [
-                "Land Bank (PHDC)",
-                "PRA",
-                None,
-                "City of Philadelphia",
-                "PRA",
-            ],
-            "sheriff_sale": ["Y", "N", "N", "N", "Y"],
-            "market_value": [2000, 500, 800, 5000, 1000],
-            "all_violations_past_year": [5, 0, 2, 0, None, 1],
-            "sale_date": ["2020-01-01", "2021-03-10", "2024-12-30", "2004-09-10", None],
-        }
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-
-        conservatorship_feature_layer = conservatorship(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["conservatorship"] = ["No", "No", "Yes", "Yes", "No"]
-
-        assert expected_gdf.equals(conservatorship_feature_layer.gdf)
-
-    def test_priority_level(self):
-        pass
-        """
-        Test the functionality of the priority_level service that generates a new column for the 
-        dataset on the basis of some branching logic in existing column values in the corresponding row 
-        (see priority_level function for more details).
-        """
-        data = {
-            "gun_crimes_density_zscore": [-0.5, 1.1, 0.5, 2.1, 0.4],
-            "all_violations_past_year": [3, 2, 0, 0, 5],
-            "l_and_i_complaints_density_zscore": [0, 0.1, 2.1, -1, 0.3],
-            "tree_canopy_gap": [0.1, 0, 0.4, 0.5, 0.2],
-            "phs_care_program": [
-                None,
-                "Trees for Watersheds",
-                "Rain Check",
-                "Trees for Watersheds",
-                None,
-            ],
-        }
-
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-
-        priority_level_feature_layer = priority_level(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["priority_level"] = [
-            "Low",
-            "High",
-            "Medium",
-            "High",
-            "Medium",
-        ]
-
-        assert expected_gdf.equals(priority_level_feature_layer.gdf)
-
-    def test_access_process(self):
-        """
-        Test the functionality of the access_process service that generates a new column for the dataset
-        on the basis of existing values in the corresponding row for city_owner_agency and market_value.
-        """
-        # Create test data
-        data = {
-            "city_owner_agency": [
-                "Land Bank (PHDC)",
-                "PRA",
-                None,
-                "City of Philadelphia",
-                None,
-            ],
-            "market_value": [1500, 500, 2000, None, None],
-        }
-        gdf = gpd.GeoDataFrame(data)
-        feature_layer = FeatureLayer(name="test", gdf=gdf)
-
-        access_process_feature_layer = access_process(feature_layer)
-        expected_gdf = gdf.copy()
-        expected_gdf["access_process"] = [
-            "Go through Land Bank",
-            "Do Nothing",
-            "Private Land Use Agreement",
-            "Buy Property",
-            "Buy Property",
-        ]
-
-        assert expected_gdf.equals(access_process_feature_layer.gdf)
 
 
 if __name__ == "__main__":
